+++ pyanalyze/node_visitor.py
@@ -15,6 +15,7 @@ from dataclasses import dataclass
 from enum import Enum
 import qcore
 import cProfile
+import json
 import logging
 import os
 import os.path
@@ -25,7 +26,7 @@ import tempfile
 import builtins
 from builtins import print as real_print
 from types import ModuleType
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 from typing import (
     Any,
     Dict,
@@ -38,7 +39,6 @@ from typing import (
     Tuple,
     Type,
     Union,
-    cast,
 )
 
 from . import analysis_lib
@@ -128,13 +128,16 @@ class FileNotFoundError(Exception):
     pass
 
 
-class Failure(TypedDict, total=False):
+class Failure(TypedDict):
     description: str
     filename: str
-    code: Enum
-    lineno: int
-    context: str
-    message: str
+    absolute_filename: str
+    code: NotRequired[Enum]
+    lineno: NotRequired[int]
+    col_offset: NotRequired[int]
+    context: NotRequired[str]
+    message: NotRequired[str]
+    extra_metadata: NotRequired[Dict[str, Any]]
 
 
 class BaseNodeVisitor(ast.NodeVisitor):
@@ -360,6 +363,7 @@ class BaseNodeVisitor(ast.NodeVisitor):
         else:
             kwargs = dict(args.__dict__)
         markdown_output = kwargs.pop("markdown_output", None)
+        json_output = kwargs.pop("json_output", None)
 
         verbose = kwargs.pop("verbose", 0)
         if verbose == 0 or verbose is None:
@@ -395,7 +399,15 @@ class BaseNodeVisitor(ast.NodeVisitor):
             failures = cls._run(**kwargs)
             if markdown_output is not None and failures:
                 cls._write_markdown_report(markdown_output, failures)
+            if json_output is not None and failures:
+                cls._write_json_report(json_output, failures)
         return 1 if failures else 0
+
+    @classmethod
+    def _write_json_report(cls, output_file: str, failures: List[Failure]) -> None:
+        failures = [_make_serializable(failure) for failure in failures]
+        with open(output_file, "w") as f:
+            json.dump(failures, f)
 
     @classmethod
     def _write_markdown_report(cls, output_file: str, failures: List[Failure]) -> None:
@@ -409,7 +421,6 @@ class BaseNodeVisitor(ast.NodeVisitor):
 
         with open(output_file, "w") as f:
             f.write("%d total failures in %d files\n\n" % (len(failures), len(by_file)))
-
             for filename, file_failures in sorted(by_file.items()):
                 if filename != UNUSED_OBJECT_FILENAME:
                     filename = filename[len(prefix) :]
@@ -525,6 +536,7 @@ class BaseNodeVisitor(ast.NodeVisitor):
         ignore_comment: str = IGNORE_COMMENT,
         detail: Optional[str] = None,
         save: bool = True,
+        extra_metadata: Optional[Dict[str, Any]] = None,
     ) -> Optional[Failure]:
         """Shows an error associated with this node.
 
@@ -540,6 +552,9 @@ class BaseNodeVisitor(ast.NodeVisitor):
           file-level ignore comments.)
         - ignore_comment: Comment that can be used to ignore this error. (By default, this
           is "# static analysis: ignore".)
+        - detail: extra detail to append to the error on a separate line
+        - save: if False, do not add the failure to the all_failures list
+        - extra_metadata: if given, is added to JSON failures output
 
         """
         if self.caught_errors is not None:
@@ -552,6 +567,8 @@ class BaseNodeVisitor(ast.NodeVisitor):
                     "obey_ignore": obey_ignore,
                     "ignore_comment": ignore_comment,
                     "detail": detail,
+                    "save": save,
+                    "extra_metadata": extra_metadata,
                 }
             )
             return None
@@ -582,8 +599,13 @@ class BaseNodeVisitor(ast.NodeVisitor):
         else:
             lineno = col_offset = None
 
-        # https://github.com/quora/pyanalyze/issues/112
-        error = cast(Failure, {"description": str(e), "filename": self.filename})
+        error: Failure = {
+            "description": str(e),
+            "filename": self.filename,
+            "absolute_filename": os.path.abspath(self.filename),
+        }
+        if extra_metadata is not None:
+            error["extra_metadata"] = extra_metadata
         message = f"\n{e}"
         if error_code is not None:
             error["code"] = error_code
@@ -595,6 +617,8 @@ class BaseNodeVisitor(ast.NodeVisitor):
             message += f"\nIn {self.filename} at line {lineno}\n"
         else:
             message += f"\n In {self.filename}"
+        if col_offset is not None:
+            error["col_offset"] = col_offset
         lines = self._lines()
 
         if obey_ignore and lineno is not None:
@@ -843,6 +867,13 @@ class BaseNodeVisitor(ast.NodeVisitor):
             help=(
                 "Write errors to this file in markdown format. "
                 "Suitable for summarizing and tracking errors."
+            ),
+        )
+        parser.add_argument(
+            "--json-output",
+            help=(
+                "Write errors to this file in JSON format. "
+                "Suitable for integrating with other tools."
             ),
         )
         parser.add_argument(
@@ -1097,3 +1128,10 @@ class _Profile(object):
         self.filename = tempfile.mktemp()
         self.prof.dump_stats(self.filename)
         print("profiler output saved as {}".format(self.filename))
+
+
+def _make_serializable(failure: Failure) -> Dict[str, Any]:
+    result = dict(failure)
+    if "code" in failure:
+        result["code"] = failure["code"].name
+    return result