+++ pyanalyze/suggested_type.py
@@ -6,13 +6,15 @@ Suggest types for untyped code.
 import ast
 from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import Dict, Iterator, List, Mapping, Sequence, Union
+from types import FunctionType
+from typing import Any, Dict, Iterator, List, Mapping, Optional, Sequence, Tuple, Union
 
-from pyanalyze.safe import safe_isinstance
+from pyanalyze.safe import safe_getattr, safe_isinstance
 
 from .error_code import ErrorCode
 from .node_visitor import Failure
 from .value import (
+    NO_RETURN_VALUE,
     AnnotatedValue,
     AnySource,
     AnyValue,
@@ -28,6 +30,7 @@ from .value import (
     MultiValuedValue,
     VariableNameValue,
     replace_known_sequence_value,
+    stringify_object,
     unite_values,
 )
 from .reexport import ErrorContext
@@ -58,16 +61,20 @@ class CallableData:
             all_values = [v for v in all_values if not isinstance(v, AnyValue)]
             if not all_values:
                 continue
-            suggested = display_suggested_type(unite_values(*all_values))
+            suggested = unite_values(*all_values)
+            if not should_suggest_type(suggested):
+                continue
+            detail, metadata = display_suggested_type(suggested)
             failure = self.ctx.show_error(
                 param,
                 f"Suggested type for parameter {param.arg}",
                 ErrorCode.suggested_parameter_type,
-                detail=suggested,
+                detail=detail,
                 # Otherwise we record it twice in tests. We should ultimately
                 # refactor error tracking to make it less hacky for things that
                 # show errors outside of files.
                 save=False,
+                extra_metadata=metadata,
             )
             if failure is not None:
                 yield failure
@@ -100,13 +107,50 @@ class CallableTracker:
         return failures
 
 
-def display_suggested_type(value: Value) -> str:
+def display_suggested_type(value: Value) -> Tuple[str, Optional[Dict[str, Any]]]:
     value = prepare_type(value)
     if isinstance(value, MultiValuedValue) and value.vals:
         cae = CanAssignError("Union", [CanAssignError(str(val)) for val in value.vals])
     else:
         cae = CanAssignError(str(value))
-    return str(cae)
+    # If the type is simple enough, add extra_metadata for autotyping to apply.
+    if isinstance(value, TypedValue) and type(value) is TypedValue:
+        # For now, only for exactly TypedValue
+        if value.typ is FunctionType:
+            # It will end up suggesting builtins.function, which doesn't
+            # exist, and we should be using a Callable type instead anyway.
+            metadata = None
+        else:
+            suggested_type = stringify_object(value.typ)
+            imports = []
+            if isinstance(value.typ, str):
+                if "." in value.typ:
+                    imports.append(value.typ)
+            elif safe_getattr(value.typ, "__module__", None) != "builtins":
+                imports.append(suggested_type.split(".")[0])
+            metadata = {"suggested_type": suggested_type, "imports": imports}
+    else:
+        metadata = None
+    return str(cae), metadata
+
+
+def should_suggest_type(value: Value) -> bool:
+    # Literal[<some function>] isn't useful. In the future we should suggest a
+    # Callable type.
+    if isinstance(value, KnownValue) and isinstance(value.val, FunctionType):
+        return False
+    # These generally aren't useful.
+    if isinstance(value, TypedValue) and value.typ in (FunctionType, type):
+        return False
+    if isinstance(value, AnyValue):
+        return False
+    if isinstance(value, MultiValuedValue) and len(value.vals) > 5:
+        # Big unions probably aren't useful
+        return False
+    # We emptied out a Union
+    if value is NO_RETURN_VALUE:
+        return False
+    return True
 
 
 def prepare_type(value: Value) -> Value:
@@ -128,8 +172,10 @@ def prepare_type(value: Value) -> Value:
     elif isinstance(value, VariableNameValue):
         return AnyValue(AnySource.unannotated)
     elif isinstance(value, KnownValue):
-        if value.val is None or safe_isinstance(value.val, type):
+        if value.val is None:
             return value
+        elif safe_isinstance(value.val, type):
+            return SubclassValue(TypedValue(value.val))
         elif callable(value.val):
             return value  # TODO get the signature instead and return a CallableValue?
         value = replace_known_sequence_value(value)
@@ -139,22 +185,27 @@ def prepare_type(value: Value) -> Value:
             return prepare_type(value)
     elif isinstance(value, MultiValuedValue):
         vals = [prepare_type(subval) for subval in value.vals]
-        type_literals = [
-            v
-            for v in vals
-            if isinstance(v, KnownValue) and safe_isinstance(v.val, type)
-        ]
-        if len(type_literals) > 1:
-            types = [v.val for v in type_literals if isinstance(v.val, type)]
-            shared_type = get_shared_type(types)
-            type_val = SubclassValue(TypedValue(shared_type))
-            others = [
-                v
-                for v in vals
-                if not isinstance(v, KnownValue) or not safe_isinstance(v.val, type)
-            ]
-            return unite_values(type_val, *others)
-        return unite_values(*vals)
+        # Throw out Anys
+        vals = [val for val in vals if not isinstance(val, AnyValue)]
+        type_literals: List[Tuple[Value, type]] = []
+        rest: List[Value] = []
+        for subval in vals:
+            if (
+                isinstance(subval, SubclassValue)
+                and isinstance(subval.typ, TypedValue)
+                and safe_isinstance(subval.typ.typ, type)
+            ):
+                type_literals.append((subval, subval.typ.typ))
+            else:
+                rest.append(subval)
+        if type_literals:
+            shared_type = get_shared_type([typ for _, typ in type_literals])
+            if shared_type is object:
+                type_val = TypedValue(type)
+            else:
+                type_val = SubclassValue(TypedValue(shared_type))
+            return unite_values(type_val, *rest)
+        return unite_values(*[v for v, _ in type_literals], *rest)
     else:
         return value
 
@@ -169,10 +220,14 @@ def get_shared_type(types: Sequence[type]) -> type:
     assert False, "should at least have found object"
 
 
-def _extract_params(node: FunctionNode) -> Iterator[ast.arg]:
+# We exclude *args and **kwargs by default because it's not currently possible
+# to give useful types for them.
+def _extract_params(
+    node: FunctionNode, *, include_var: bool = False
+) -> Iterator[ast.arg]:
     yield from node.args.args
-    if node.args.vararg is not None:
+    if include_var and node.args.vararg is not None:
         yield node.args.vararg
     yield from node.args.kwonlyargs
-    if node.args.kwarg is not None:
+    if include_var and node.args.kwarg is not None:
         yield node.args.kwarg